Fixes everything except DropDownView
All that's left is to get DropDownView into the utilities
