+++ orienteer/bot/cogs/ytpromo.py
@@ -3,6 +3,7 @@ from disnake.ext import commands
 from disnake.ext.commands import Bot, BucketType
 from orienteer.bot.calls import ytpromo
 from orienteer.bot.utils import embeds
+# from orienteer.bot.utils.ytpromo import DropDownView
 from orienteer.bot.utils.content_locale import Errors
 from orienteer.general.data.orienteer.database import database_helper
 from orienteer.general.utils.dtos import UserDTO
